services/market-data/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.23.0
redis==4.3.4
python-dotenv==0.19.0 